done JDBC

+++ src/main/java/com/amigoscode/customer/CustomerDao.java
@@ -6,14 +6,14 @@ import java.util.Optional;
 public interface CustomerDao {
     List<Customer> selectAllCustomers();
 
-    Optional<Customer> selectCustomerById(Integer id);
+    Optional<Customer> selectCustomerById(Long id);
 
     void insertCustomer(Customer customer);
 
     boolean existsPersonWithEmail(String email);
 
-    boolean existsPersonWithId(Integer id);
+    boolean existsPersonWithId(Long id);
 
-    void deleteCustomer(Integer customerId);
-    void updateCustomer(Integer customerId, Customer customer);
+    void deleteCustomer(Long customerId);
+    void updateCustomer(Long customerId,Customer customer);
 }
